[refactor] :recycle: add sufix Helper to utils modules (#276)
