+++ scripts/subscribe_wallets_to_push_channel.ts
@@ -9,8 +9,8 @@ import * as PushAPI from '@pushprotocol/restapi';
 import { ENV } from '@pushprotocol/restapi/src/lib/constants';
 
 import { IUser } from '../src/models/user';
-import { Logger } from '../src/utils/logger';
-import { generatePrivateKey } from '../src/utils/keyGenerator';
+import { Logger } from '../src/utils/loggerHelper';
+import { generatePrivateKey } from '../src/utils/keyGeneratorHelper';
 
 dotenv.config();
