+++ src/config/database.ts
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose';
 
-import { Logger } from '../utils/logger';
+import { Logger } from '../utils/loggerHelper';
 import { MONGO_URI } from '../constants/environment';
 
 /**
