+++ src/controllers/ImageController.ts
@@ -1,8 +1,8 @@
 import { FastifyReply, FastifyRequest, RouteHandlerMethod } from 'fastify';
 
 import { User } from '../models/user';
-import { Logger } from '../utils/logger';
-import { returnErrorResponse, returnSuccessResponse } from '../utils/responseFormatter';
+import { Logger } from '../utils/loggerHelper';
+import { returnErrorResponse, returnSuccessResponse } from '../utils/responseFormatterHelper';
 import { uploadToICP, uploadToIpfs, downloadAndProcessImage } from '../services/uploadService';
 
 interface UploadBody {
