+++ src/services/externalDepositsService.ts
@@ -1,7 +1,7 @@
 import { gql, request } from 'graphql-request';
 
 import { User } from '../models/user';
-import { Logger } from '../utils/logger';
+import { Logger } from '../utils/loggerHelper';
 import Transaction from '../models/transaction';
 import { sendTransferNotification } from './notificationService';
 import { LastProcessedBlock } from '../models/lastProcessedBlock';
