+++ src/services/paymasterService.ts
@@ -1,8 +1,8 @@
 import { ethers } from 'ethers';
 
-import { Logger } from '../utils/logger';
-import { createPaymasterAndData } from '../utils/paymaster';
+import { Logger } from '../utils/loggerHelper';
 import { PackedUserOperationType } from '../types/userOperation';
+import { createPaymasterAndData } from '../utils/paymasterHelper';
 import { PAYMASTER_MIN_BALANCE, PAYMASTER_TARGET_BALANCE } from '../constants/environment';
 
 /**
