+++ src/services/userService.ts
@@ -1,5 +1,5 @@
-import { Logger } from '../utils/logger';
 import { User, IUser } from '../models/user';
+import { Logger } from '../utils/loggerHelper';
 import { SETTINGS_NOTIFICATION_LANGUAGE_DFAULT } from '../constants/environment';
 import { ComputedAddress, computeProxyAddressFromPhone } from './predictWalletService';
 import { subscribeToPushChannel, sendWalletCreationNotification } from './notificationService';
