+++ src/utils/dynamicGasHelper.ts
@@ -1,6 +1,5 @@
-import { Contract, BigNumber } from 'ethers';
-
-import { Logger } from './logger';
+import { BigNumber, Contract } from "ethers";
+import { Logger } from "./loggerHelper";
 
 /**
  * Get gas limit for a transaction w/ dynamic gas.
