+++ src/utils/keyGeneratorHelper.ts
