--- src/utils/paramsUtils.ts
@@ -1,4 +1,0 @@
-export const isValidUrl = (url: string): boolean => {
-  const urlPattern = /^(https?:\/\/)?([\w-]+\.)+[\w-]+(\/[\w- ./?%&=]*)?$/;
-  return urlPattern.test(url);
-};
